# frinq-mobile/docs/openapi-snapshot.json
{
  "components": {
    "schemas": {
...
      "AccountState": {
        "properties": {
          "banned": {
            "title": "Banned",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "onboarding_state": {
            "title": "Onboarding State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "onboarding_state",
          "banned"
        ],
        "title": "AccountState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PriorSession": {
        "properties": {
          "answers": {
            "title": "Answers",
            "type": "object"
          },
          "is_complete": {
            "default": false,
            "title": "Is Complete",
            "type": "boolean"
          },
          "last_page": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Page"
          },
          "status": {
            "default": "pending",
            "title": "Status",
            "type": "string"
          },
          "submission_id": {
            "title": "Submission Id",
            "type": "string"
          }
        },
        "required": [
          "submission_id",
          "answers"
        ],
        "title": "PriorSession",
        "type": "object"
      },
...
      "SendOTPRequest": {
        "properties": {
          "phone": {
            "title": "Phone",
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "title": "SendOTPRequest",
        "type": "object"
      },
      "SendOTPResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "SendOTPResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyOTPRequest": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "phone": {
            "title": "Phone",
            "type": "string"
          },
          "platform": {
            "enum": [
              "ios",
              "android",
              "web"
            ],
            "title": "Platform",
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code",
          "platform"
        ],
        "title": "VerifyOTPRequest",
        "type": "object"
      },
      "VerifyOTPResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "prior_session": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PriorSession"
              },
              {
                "type": "null"
              }
            ]
          },
          "refresh_token": {
            "title": "Refresh Token",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AccountState"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "user"
        ],
        "title": "VerifyOTPResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Frinq Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/otp/send": {
      "post": {
        "operationId": "send_otp_route_api_v1_otp_send_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendOTPResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Send Otp Route",
        "tags": [
          "otp"
        ]
      }
    },
    "/api/v1/otp/verify": {
      "post": {
        "description": "Verify the OTP code with Twilio. Then (only on success) look up\nprior session. Sequential \u2014 was previously parallel via create_task but\nthe concurrency was making failure modes unpredictable.",
        "operationId": "verify_otp_route_api_v1_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOTPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyOTPResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Otp Route",
        "tags": [
          "otp"
        ]
      }
    },
...
  }
}